Fix graceful shutdown problem
The s3 client shutdown caused ASan memory leak sanitizer to trigger
the error message. This was caused by the way tls session is implemented
in seastar. The ss::tls::session::session_ref d-tor schedules
asynchronous close operation without waiting (since the path is
futurized). The close op is scheduled using with_timeout and
expected to run in the background. It's not a problem if the app
is still running but can cause memory leak if the application stops
before background op is completed.

The close op is relatively lengthy since it have to send the
close_notify message to the server.