archival: Add archival scheduler service
The service is responsible for S3 uploads. It wires up with log_manager
and partition_manager. Periodically, it checks what partitions are
available and adds them to the queue. It fetches a bunch of sealed
segents from the queue periodically and uploads them. Only leader for
a particular NTP uploads segments that belongs to that NTP.

The segments are deleted from S3 when corresponding segment is deleted
locally. Algorithms for selecting both upload and delete candidates are
supposed to be configurable. In this PR only one policy for upload and one
for delete is implemented. For uploads, we select only non-compacted
segments. For deletes, we delete segment from S3 as soon as possible after
corresponding segment is deleted locally. There're extension points so
new policies could be added in future PRs.

The period used for partition reconciliation and upload can be configured.
Number of simultaneous uploads is also limited by configuration.

Currently, garbage collection is left out. It will be added in future
PRs.