archival: Add cloud_storage module
Add manifest impelemntation. The implementation is the same as in
archival module.

The cloud_storage module is supposed to be shared between the
archival and storage modules. It will have all code responsible for
data uploading and downloading to and from S3. It's also responsible
form manifest formats and naming schema for S3 (object name generation).

The archival module will be responsible for scheduling uploads and
reconcililng the ntp's.