utils: Add retry_chain_node class
Retry chain is a bunch of fiber that are working together and share the
same deadline. The fibers can retry on error. The retry_chain_node adds
ability to generate correct backoff interval durations (with respect to
common deadline time).
Also, the retry_chain_node can be used for cancelation. If abort_source
is passed to the c-tor it can be used by any fiber. The abort_source is
checked automatically when the fiber calls retry_chain_node::retry.
The retry_chain_node can also be used to generate unique fiber ids. The
format of the id is [fiberXX~YY~...|R|Tms] where XX and YY etc are node
ids. R is a number of retries being made. T is current time budget (the
time before the deadline).