archival: Use dirty offset of the log segment
Archival upload algorithm can fail if the dirty offset of the rolled log
segment is greater than committed offset (as described in #1521). The
committed offset is used as a last uploaded offset in the manifest and
the search for the next upload candidate is started from this offset.
Then the uploaded segment becomes the next uploaded candidate once again
(this time partial upload is triggered with length 0).

This commit retargets archival subsystem from committed offset to dirty
offset (which is always consistent with the actual file layout on
dicsk). The archival service uses partition manager to read high
watermark offset of the partition. Then it compares the dirty offset
with high watermark and decides if the segment can be uploaded. It never
uploads a segment that contains an offset which is higher than high
watermark.

This change makes archival service tolerant to a situation when the
committed offset is inconsistent (as described in #1800).

This change might introduce the following problem: when the log is
truncated on a leader the already uploaded log segment might be
truncated as well (since we upload by dirty offset). But check against
high watermark should prevent this situation.