cloud_storage: Split initialization of the remote
Previously, the remote was initialized inside the archival::service
instance. This is problematic since it has to be shared between the
partition_recovery_manager, the archival::service, and the shadow
indexing services (in the future).

This commit splits the probe inside the archival::service into two
components. The first one is created inside the cloud_storage::remote.
And the second one is located inside the service. The metrics
registration is done by each subsystem individually.

Also, the configuration is split into to components. The part which is
used by the archival::service left in place. The part which is used
to initialize the cloud_storage::remote moved to the new place.

In the application code the cloud_storage::remote is created as a
sharded<> instance and passed to the archival::service instance upon
creation. We need this to be able to pass the remote to other
subsystems.

Last but not least, the download_segment method of the remote now
accepts manifest::key instead of segment_name to allow using
both remote_segment_path and segment_name (since both can be keys in
manifest).