cloud_storage: Add offset translation layer
Implements offset translator. The component uses the manifest as a
data source to translate from raft offsets to kafka offsets on the fly.
It implements streaming interface and can patch segments on the fly
as the stream is being scanned.

The component can also adjust segment file names.

The offset_translator can only work with existing offset boundaries.
This means that we can only lookup a segment name wich is present in
the manifest. The streaming feature also depend on segment boundaries.
The start of the stream shuld be aligned with the segment meta in the
manifest.

The object is used to create patched segments with all raft
configuration batches removed and offsets and CRCs adjusted.

It will also be reused in shadow indexing.