cluster: Integrate partition_recovery_manager
with cluster::partition_manager. Add extra toipc configuration
option 'x-redpanda-recovery' which will be used to enable S3
recovery.

Use remote to initialize partition_recovery_manager (if the
feature is enabled) in application.cc and pass
partition_recovery_manager to the partition_manager on startup.
The partition_reocovery_manager may not be initialized if the
cloud_storage is disabled.

The behavior of the partition_manager is the same. The
partition_recovery_manager is not invoked.