cloud_storage: Limit number of materialized segments
Limit number of segments that can be in the materialized state at the
same time to one segment. Trigger GC when the materialization of the
segment is needed.

The following cases are possible when new segment is accessed:
1. no other readers are accessing other segments at the moment
2. other readers are actively reading from other materialized segments

In the first case all other materialized segments will be deleted
immediately. In the second case the segments that are used at the moment
won't be deleted immediately but only after some period of time. When
all readers will stop using the segments.

So in the worst case we will hold as many materialized segments as we
have active kafka fetch requests.