cloud_storage: Use segment_name_components struct
Use segment_name_components instead of the segment_name in the manifest
and some interfaces. The step is required to limit memory use and
fragmentation.

The manifest object contains a map<segment_name, segment_meta>
datastructure. The key is a string that contains a segment name. The
problem with this approach is that segment names can grow larger than
the SSO limit (15 characters). After that the keys of the map object are
allocated on heap.

This commit changes the type of the underlying key from sstring to
segment_name_components struct (which contains base offset and term).

Important: manifest doesn't provide pointer stability guarantee
anymore.