cloud_storage: Optimize memory use in SI
The remote_partition stores an object for every log segment which was
uploaded to S3 and is available through it. The object is either a
offloaded_segment_state or materialized_segment_state. The first one
contains an object key in S3. The second one contains remote_segment and
a collection of readers. Both objects are allocated on a heap.

The offloaded_segment_state is a small struct that contained a pointer to
the heap allocated data (the object key is a string). The object key
could be heap allocated when it's too large for SSO. The
materialized_segment_state is a large object with a lot of fields which
are also reference some other heap allocated objects. The collection
that maps from offsets to state objects was an std::map. The key of the
map is an offset and the value is a variant that can be either offloaded
or materialized segment. To make this variant small both structs are
heap allocated.

This design creates a following problem. The number of log segments in
the S3 bucket can be large (much large than the number of segments on
disk). And even if SI is not actively used we're allocating memory for a
std::map node, and also we're allocating 24 bytes for
offloaded_segment_state (which in turn contains pointer to heap
allocated structure which is 100+ bytes on its own).

This commit introduces the following changes:
- The object key inside both offloaded_segment_state and
  materialized_segment_state is now a base_offset instead of segment
  file name.
- The manifest object now supports lookups based on base-offset of the
  segment. This allow us to use an offset as a key.
- The variant type contains offloaded_segment_state instead of the
  pointer. This eliminates all memory allocations if the remote segment
  is not actively used. So the variant type stores either offset or a
  pointer to materialized_segment_state.
- The data structure is swithced to 'absl::btree_map'.