cloud_storage: Add an offset_index component
Also the encoder and decoder for the delta-FoR compression schema.
The index stores three types of offsets: redpanda offsets, kafka
offsets, and file offsets. You can do a linear search on any one of them
and return all three values that correspond to the same record batch.

This index is supposed to be used to eliminate full segment scans on
pont queries. The remote_segment is designed with batch reads in mind.
It can only be scanned from the begining, otherwise we can't reconstruct
offset translation data. The index will allow us to start scanning the
segment from any indexed offset.