rptest: Add topic recovery end to end test
The test uses FranzGoVerifiableProducer and
FranzGoVerifiableSeqConsumer. The test configures redpanda to use idle
timeout for S3 uploads. It produces the data then it stops redpanda
nodes and removes data from data directories of the nodes. After that it
restarts the nodes and runs topic-recovery. When the topic recovery is
invoked it starts verifable consumer.

One of the test variants sets retention to very low value on a recovery
topic and enables SI. This forces verifer to use SI instead of the local
data. This allows us to check that SI works after topic recovery.