cluster: Initialize archival STM snapshot
Initialize archival STM snapshot after running topic recovery.
Previously the code used 'add_segments' method of the
'archival_metadata_stm' right after topic recovery. The problem is that
this method needs to replciate the configuration batch to succeed. This
can't happen because at this point we don't have a leader yet. Also,
since we're initializing it this way STM doesn't try to download the
manifest on its own.

The fix here constructs the snapshot on disk from the manifest.