cloud_storage: Add access_time_tracker
The component can be used to track file access times using
probabilistic data structure. Internally it uses CM sketch that
tracks timestamps instead of counters. The idea behind this is
that instead of incrementing counters in CM-sketch one can write
a timestamp and apply the same logic.

When we insert the timestamp we put it into N cells of the CM-sketch.
And when the value needs to be extracted we can read from N cells and
find the min value among them. This will give us the resulting timestamp.

There is a room for conflicts. In case of conflict 1 to N cells will be
overwritten by subsequent updates. These inserts are guaranteed to use
larger timestamps. If less then N cells were overwritten we're
guaranteed to read correct value by selecting smallest timestamp. If all N
cells were overwritten due to conflicts we will read larger timestamp.

The intended use case is access time tracking in cache. For this use
case it's important to not underestimate access times since we don't
want to evict hot item from the cache.

The data structure is fixed size and lock free. Essentially, this is a
hash table of fixed size with atomic integers (with padding) as values.
In shadow indexing cache we know in advance how many items it's expected
to hold so we can size the CM sketch properly.

The data structure will be shared between the shards. Since every load
or store only touches few random elements it's unlikely that we will
introduce contention. Especially given that every shard uses it's own
set of keys.