archival: Do not run upload/sync/housekeeping loops ...
util we have stable leadership. The code was using is_elected_leader. In
this state the underlying raft group is not guaranteed to be able to
replicate the message. All this internal background loops are actually
replicating messages and can fail if the leadership is not stable.

The commit fixes this by using 'cluster::partition::is_leader' instaead
of using 'cluster::partition::is_elected_leader'.