archival: Add upload_housekeeping_service
The service is not doing anything yet. This is only a scaffolding for
the next commits. It contains the intrusive list of all ntp_archiver
instances. The instances are registered by the cloud::partition after
being created. The upload_housekeeping_service is created in
application.cc as a sharded service and passed to the partition_manager
and further to the partition.