archival: Upload housekeeping service
Implement scheduling logic for upload housekeeping service. The service
is not tied to ntp-archiver anymore. It schedules housekeeping_job
objects. The ntp_archiver is supposed to register its periodic job which
will be executed by the service.

The service runs all jobs once per epoch. The epoch is one
cloud_storage_housekeeping_interval_ms period. If the cloud storage is
idle for cloud_storage_idle_timeout_ms the housekeeping may be triggered
earlier. But in this case it may also be put on pause if the cloud
storage will become active again. The goal is to prevent housekeeping to
run in bursts and run it more evently if possible.