archival: Create housekeeping job in ntp_archiver
The housekeeping job is a local_adjacent_segment_merger which uploads
segments from local disk to S3 in case if there are a lot of small
segments.

The 'ntp_archiver' is extended. The method that returns next upload
candidate based on predicate is added. The new 'upload' method can be
used to upload this upload cadidate to the cloud storage. Both methods
are supposed to be used externally, by the upload_housekeeping_serivce.
The 'upload' method is taking hold of the semaphore that protects the
upload path.

The 'upload', 'upload_tx, and 'upload_segment' methods are now accept
the 'retry_chain_node' optinally. By default they're using internal
retry chain of the 'ntp_archiver'. But when
'upload_housekeeping_service' is calling 'upload' method it has to pass
its own 'retry_chain_node' to supress notification that the upload may
trigger. Otherwise the first upload done by the
'upload_housekeeping_service' will send notification back to
housekeeping service and will put subsequent uploads on pause.

The 'find_upload_candidates' first uses the predicate to find candidates
in the manifest. Then it uses 'segment_collector' to create a
multi-segment upload candidate with locks. The
'local_adjacent_segment_merger' uses predicate that scans manifest for
the next series of small segments that can be merged and re-upload. The
merging is not happenes on disk, only in the cloud storage.