archival: Add can_update_archival_metadata method
Replace three methods with the same implementation:
upload_loop_can_continue, sync_manifest_loop_can_continue, and
housekeeping_can_continue.

The reason behind this is that the implementation is the same and
it could be confusing to have three different methods for all cases
where we are updating the metadata.