cloud_storage: Add synchronization to manifest
The partition_manifest is updated by the archival_metadata_stm
asynchronously. With manifest serialization running concurrently this
creates a problem. To avoid this the semaphore is added to the
partition_manifest. It is acquired in the 'apply' method of the archival
STM and when manifest serialization is invoked.

This mechanism is temporary and won't be needed when manifest
compression will be merged (since column store is transactional and
supports shallow copy).