storage: remove memory log implementation
The storage fixture test flipped a coin to select disk or memory and
cache or not cache for the backend configuration. The same coin flip was
used in both cases such that disk always had a configured cache. It's
unknown if this was intentional or not, but preserving this behavior in
any case.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>