cloud_storage: Add spillover end to end test
The test produces records and rolls segmens frequently to populate the
manifest. Then the test calls `apply_spillover` on archiver and
validates uploaded content and consumes via the kafka layer.

The test starts with current start offset instead of
archive_start_offset. When the read path will be updated the tests
should also be updated to start with archive_start_offset.