cluster: Improve archival_metadata_stm::spillover_cmd
Currently, the command has only one field that contains the new start
offset after spillover. The STM applies this offset and computes the
metadata for the spillover manifest based on its state (the manifest
itself). This commit modifies the command to store the metadata itself.
With metadata in the command the STM can validate the command and only
apply it if it's safe (no data loss).