archival: Add async_data_uploader class
The async_data_uploader can be used to provide the
ss::input_stream<char> that cover certain offset range and generate
metadata for it. The class makes a storage::reader and uses it to go
through the offset range while serializing everything into disk format.
The reader is bypassing both batch cache and readers cache to avoid
interference with kafka fetch requests and avoid cache pollution.