archival: Add size-limited mode to segment_upload
Currently, segment_upload has to know first and last offsets of the
uploaded segment to create an upload. This works in situations when we
know the last offset (we always know the first offset of the upload). In
some situation we can't know the last offset in advance. For instance,
if we need to make several sequential uploads (if the partition is
lagging). In this case we need to make first upload with certain size
and use its last offset to start next upload with certain size.

This PR adds a new mode that takes start offset and size range (man/max
size) and trires to create an upload that starts at specific offset and
fists inside the size range. The sizing is not precise and the actual
size can overshoot by 32KiB. In this mode the segment is always ended on
segment or index boundary.

The tests are only checking normal (non-compacted) segments because we
don't need this mode for compacted segments. Compacted segments are
always reuploaded so we alwyas know first/last offset of the upload.