utils: Implement auto_fmt<> template
The template uses CRTP pattern to implement formatting for small
structures. The structure has to be derived from it first. You can
provide the field delimiter as template non-type parameter:

  struct foo : auto_fmt<foo, ','> {...};

After that the values of the struct could be formatted using << operator
or libfmt. The output is a list of all fields separated by the
delimiter. The default delimiter is space character.