archival: Add fixture-tests for archiver_service
The tests are running a cluster fixture with 2 shards on every cluster
node. There are several test cases that test behavior of the
archiver_service after partition movement, cross shard movement and
leadership transfer.

There is also a "torture" test that invokes all these operations
randomly and checks invariant afterwards. This commit also adds
extensive fixture and makes some changes to redpanda fixture and cluster
fixture. The changes are needed to be able to initialize TS related
parameters in the cluster.