archival: Add archiver_workflow_api interface.
The workflow is a buliding block for the new archiver. Some workflows
that will be added in future commits are
- data upload workflow which uploads segments to S3
- housekeeping workflow which removes old data from S3
- read replica workflow which downloads the manifest and replicates it

When the partition becomes a leader the archiver_manager is supposed to
start corresponding workflows for it. The workflow is supposed to run
only while partition is a leader and only in one term.

The workflow may interact with the reset of the system thorugh two
interfaces.
- archiver_operations_api implements basic building blocks for the
  workflow e.g. segment or manifest upload.
- archiver_scheduler_api implements scheduling for different workflows.
  The workflow calls the method 'maybe_suspend_*' of the scheduler which
  returns a futur that can be suspended depending on circumstances. The
  scheduler "knows" about the data backlog size, data rate and cloud
  resource usage of the partition.

The workflow is also an extention point. New features could be added by
implementing new workflows.
This commit only adds an interface for the workflow. The actual
workflows will be added later.