archiver: Add 'archiver_data_upload_workflow' class
The workflow is used to upload partition data to the cloud storage. It's
implemented using boost MSM library as an explicit state machine.

Currently we're relying on the background fibers to create active objects
Usually, the code creates a background fiber which would run until 'stop'
method is called. The object can interact with the fiber and both of them
share the same state. For instance, the object can add a command to the
queue and expect the background fiber to extract the command, execute it
and send the result back. This can be considered an anti-pattern in some
cases for several reasons.
- Graceful shutdown becomes a problem (we usually not propagating
  exceptions from the background fiber to the 'stop' method).
- The execution is not deterministic and depends on scheduling heavily.
- The main problem is that it's difficult to unit-test the code that uses
  this pattern.
- Requires synchronization when the methods of the object has to
  communicate with the background fiber (the command queue in the example
  above).
- It's prone to liveliness issues. The background fiber can get stuck
  on something and it's not an easy feat to detect this.

In this commit the background loop is represented as an FSM explicitly.
Every asynchronous operation is running in the background as
before but the logic that connects the operations is represented using
FSM state transitions. The asynchronous operations invoked by the
workflow are available through the abstract virtual interfaces so they
could be easily swapped in tests. When the asynchronous operation
finishes the result of the operation triggers state transition of the
FSM. The result could be either the value or the error. The error could
be recoverable or fatal. The state transition functors are responsible
for managing the state (some data could be carried over to other
states).

This allows us to test the workflow independently from the API in the
unit-test (not fixture tests). Since the asynchronicity of the API calls
is confined to state transitions of the FSM the methods of the API mock
can return ready futures so we can simulate things like timeout errors.

Another benefit of the FSM approach is that we can list all possible
outcomes and all possible transitions and add individual tests for them
(which is done in the next commits).