archival: Change interface of the segment_upload
Previously, the 'segment_upload' created the input_stream during
initialization. This stream then could be returned to the client by
invoking 'detach_stream'. This is not convenient because during the
segment upload we may need to retry the upload multiple times. But with
the existing interface we will be forced to recreate the
'segment_upload' and redo the offset range search every time.

To avoid this pitfall the interface was changed. The stream is no longer
cached in the 'segment_upload' object. Instead, the 'make_input_stream'
method is exposed. The method can be used to create the stream. This
also eliminates the need to use input_stream_fanout utility because now
we can create as many streams as we need.