archival: Use archiver_operations_api
Use ops api in the ntp_archiver. Pivot the upload loop into two parts.
First is the legacy part which works the way it worked before this
change. The second part uses archiver_operations_api passed through the
c-tor. It's only used if the ntp_archiver was initialized in a specific
way. So it could be enabled or disabled.

At this point the new upload loop uses archiver_operations_api to upload
segments but it uses old logic to implement housekeeping (it's not
ciritical since the goal is to stop reading from disk in TS directly).

The upload loop throttling is implemented using traditional approach.
The upload loop computes backoff directly without taking into account
what other archivers are doing.