storage: Limit size based offset_range_size search
Use log committed offset to avoid returning dirty offsets from
find_above_size_bytes. The segment index is not always consistent with
the segment and in some cases it may contain entries for offsets which
are still 'dirty'. The segment size on the other hand only accounts for
committed part. This may break the offset range size computation.