archival: Fix segment roll in the cloud
Previously, we have always used size based upload lookup. In this mode
the upload builder searches for the upload candidate that starts from
some start offset (last uploaded offset + 1 or from 0 if the manifest
is empty) and tries to find upload candidate that has enough data. The
problem here is that the upload candidate may overshoot LSO.

To avoid this problem this commit updates the algorithm. Now the builder
is trying to find the upload candidate that starts at the start offset
nad ends at LSO-1. When the candidate is found it checks its size. If
the size is within the limits it proceeds with upload. If the size
overshoots it uses size bound search request. It's guaranteed that the
upload candidate will not overshoot in this case.