ct: Add request_balancer component
The request balancer is used to move write requests betwen shards.
Current implementation is very basic. The intention is to show how this
could be implemented. The balancing policy can be swapped but shards do
not exchange any information to make informed decisions.

The request_balancer only implements a basic data flow. The
write_request instance is extracted from the write_pipeline and then
submitted to another shard (through a proxy object). The result is
propagated back to the original write request and eventually to the
caller.

The fast path in no op. Initially, we can just use basic balancing
policy that disables actual balancing and leaves all write requests on
the original shard. In the future we need to add code that propagates
the information needed to make descisions (by utilizing existing
mechanism in the cloud_storage::remote) and use it to make more optimal
decisions.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>