kafka: Add cloud_topic_partition
The 'replicated_partition' has a lot of overhead. It is integrated with
the Tiered-Storage and takes it into account in the read and write paths
and also when computing offsets.

The 'cloud_topic_partition' is a simplified implementation of the
'partition_proxy' that doesn't "know" anything about the Tiered-Storage.
It's supposed to be created for every partition with cloud topics
property enabled. The Tiered-Storage can't be turned on for such topic
and cloud topics can't be disabled for the partition anyway.

Future commits will add calls to cloud topics to the
'cloud_topic_partition'. Currently it's just a wrapper around the
cluster::partition which translates offsets from log to kafka.

This commit also adds cloud_topics/api.h that contains the interface
for the cloud_topics::app which is needed to break the cyclic depndency.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>