ct: Dataplane refactoring/cleanup
This commit addresses several interconnected changes aimed at separating
the data plane and control plane. The current design has the read path
fetching placeholders from the partition, with knowledge of the 'cluster'
subsystem. The 'interfaces' submodule acts as an intermediary,
complicating code evolution due to potential cyclic dependencies and
lack of clear API boundaries.

This commit does few things to improve this. First, to avoid ciclyc
dependencies the 'app' class is split into the 'app' which is not uses a
PIMPL idiom, the 'api' that serves as an interface, and the 'app_impl'
that implements the 'api'. With this setup the 'kafka' layer doesn't
have to depend on the entire CT. It can reference the 'ss::sharded<app>'
instead. The 'app' depends only on 'api' and doesn't depend on any
implementation.

The 'read' and 'write' pipelines and requests are changed to avoid
converting between the 'record_batch_reader' and the collection of
'record_batch' values all the time. The 'write_and_debounce' method of
the 'write_pipeline' still accepts the 'model::record_batch_reader'. The
'read_pipeline' accepts a list of placeholder batches instead of the
log_reader. The top level that calls the read_pipeline is supposed to
fetch the placeholder batches from the control plane
(cluster::partition). Because of that the interface has to be changed.

The 'L0_read_path' was renamed into the 'read_path' and the
'L0_fetch_handler' was replaced with the 'fetch_handler' class. The idea
here is that the data plane shouldn't care about the distinction between
the L0 and L1. The 'read_pipeline' accepts the list of placeholders that
acta as coordinates for the actual data. In the future we can pass some
special 'coordinate' type instead of 'dl_placeholder' batches. The
'coordinate' should be able to point to both L0 and L1 depending on the
object. This commit doesn't make this change but prepares everything for
it. The 'coordinate' type will be added later.

Finally, because the read path doesn't have to query the partition
anymore the 'interfaces' submodule was removed. The work that the
'interfaces' submodule was doing is supposed to be done by the control
plane. The results should be passed to the data plane directly in the
'read_request' instead of being queried from the partition.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>