kafka: Add cloud_topic_partition
This commit adds a 'cloud_topic_partition' class which is an
implementation of the 'partiton_proxy'. The implementation connects the
control plane (cluster::partition and dl_stm) with the data plane
(cloud_topics::app).

The 'make_partition_proxy' function now accepts the
'sharded<cloud_topics::app>' as a parameter. Depending on the type of
the partition (CT vs normal) it creates either 'replicated_partition' or
'cloud_topic_partition'. The call sites are updated with few exceptions.
The 'sharded<cloud_topics::app>' reference is propagated to callers.

The main goal here is to separate the control plane and data plane.
There is no explicit API boundary for the control plane. The
'cluster::partition' is queried directly. But it will be possible to
introduce the explicit control plane API in the future.

The data plane API boundary is defined by the 'cloud_topics::app'
explicitly.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>