ssx: Add mutex with checkpointing mechanism
The mutex is named and works pretty much the same way our mutex works.
The difference is that every time the mutex is acquired the caller has
to pass a string that contains the name of the callsite.

If the mutex is acquired the checkpoint is created. The checkpoint is
simply a name of the callsite which was able to acquire the mutex and
the timestamp. This information is stored in the mutex until the mutex
is releasaed. When this happens the checkpoint is erased.

If any error occured the checkpoint information is included into the
exception.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>