cluster: Add 'prefix_truncate_record'
Add new library to avoid cyclic dependency. The ctp_stm needs to be able
to decode prefix_truncate_record messages but the only way to do this
now is to depend on the whole cluster library. This commit extracts the
prefix_truncate_record into a separate library from which both cluster
and ctp_stm depend.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>