ct: Add read_fanout component
The component handles vectorized read requests which are fetching from
multiple cloud storage objects. It scatters such requests so they could
be processed concurrently. Then it gather the results and propagates a
single result to the client.

Currently, such requests are handled in all or nothing manner. Partial
failures are not permitted.

Reuquests which target only one object are propagated to the next stage
as is.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>