ct: Add read_request_scheduler component
This is a very simple scheduler that redirects the request that reads
from the same L0 object to the same shard. The shard is decided based on
the hash of the L0 object name. The idea is that with stable mapping
between objects and shards we can do debouncing after scheduling so the
debouncing don't need to be performed cross-shard.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>

WIP tests