ct: Fetch byte range instead of the full object
The read path doesn't guaratnee that the reads from different partitions
will be aligned to hit the same L0 object simultaneously. This makes
fetching full objects less efficient. This commit makes read path to
fetch only the relevant byte range which will decrease the read
amplification by a lot. In memory caching is used to cache materialized
batches. This change will slightly increase the number of GET requests
that we will be using but GET requests are much cheaper compared to PUT
requests and the caching strategy that we have at the moment is not good
anyway.

In the future the caching strategy should be improved to avoid writing
all data to disk and the cache itself shold be improved to allow us to
store large number of small objects.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>